backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
pydantic
python-multipart
boto3
openpyxl